Split code for comments. text removed on success
