+++ js/delete_post.mjs
@@ -1,42 +1,3 @@
-// /**
-// * Deletes a post when the delete button is clicked.
-// * @async
-// */
-// const deleteButton = document.getElementById("delete-post-button");
-
-// if (deleteButton !== null) {
-//   deleteButton.addEventListener("click", async () => {
-//     const urlParams = new URLSearchParams(window.location.search);
-//     const postId = urlParams.get("id");
-
-//     if (postId === null) {
-//       console.error("Post ID not found in URL");
-//       return;
-//     }
-
-//     const token = localStorage.getItem("accessToken");
-
-//     const fetchOptions = {
-//       method: "DELETE",
-//       headers: {
-//         Authorization: `Bearer ${token}`,
-//       },
-//     };
-
-//     const response = await fetch(
-//       `https://api.noroff.dev/api/v1/social/posts/${postId}`,
-//       fetchOptions
-//     );
-
-//     if (response.ok) {
-//       console.log("Post deleted successfully");
-//       window.location.href = "feed.html";
-//     } else {
-//       console.error("Failed to delete post");
-//     }
-//   });
-// }
-
 /**
 * Deletes a post when the delete button is clicked.
 * @async
